requirements.txt
asgiref==3.8.1
Django==5.1.2
django-taggit==6.1.0
docx==0.2.4
lxml==5.3.0
Markdown==3.7
pillow==10.4.0
psycopg2==2.9.9
python-docx==1.1.2
python-dotenv==1.0.1
reindent==3.5.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
